
\name{diffuse}
\alias{diffuse}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{chessbase}
\description{
 This package provides a cleaned version of the MillionBase 2.2 database of Chess games published by Ed Schroder.
}
\usage{
data(chessdata) # Loads dataset
}
\arguments{

}
\details{
    }
%- maybe also 'usage' for other objects documented here.
\references{}
\author{Cody Ross and Bret Beheim}
\examples{
data(chessdata) # Loads dataset
 }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }